requirements.txt
python-telegram-bot==20.6
pycryptopay==0.1.3
Flask==2.3.3